requirements.txt
asgiref==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.6
django-import-export==2.5.0
django-pandas==0.6.4
et-xmlfile==1.0.1
jdcal==1.4.1
joblib==1.0.1
MarkupPy==1.14
numpy==1.20.1
odfpy==1.4.1
openpyxl==3.0.6
pandas==1.2.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
tablib==3.0.0
threadpoolctl==2.1.0
xlrd==2.0.1
xlwt==1.3.0
